# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz App",
        "description": "Documentación de mi API REST",
        "contact": {
            "email": "kvfa13@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Obter todos os quizzes",
                "description": "Retorna uma lista de todos os quizzes disponíveis.",
                "operationId": "9537710f5308de2caa84652ba9bc92af",
                "responses": {
                    "200": {
                        "description": "Lista de quizzes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 19
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Quiz 2"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Queremos eliminar um quiz de forma correta"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Criar um novo quiz",
                "description": "Cria um novo quiz com base nos dados fornecidos. Apenas usuários com o papel de administrador podem criar um novo registro.",
                "operationId": "8e9f7aa63a127de0bd41e7f4bbcac67b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Quiz 5"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Esto es una descripcion 5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quiz criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Quiz 5"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Esto es uma descrição 5"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T02:19:50.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T02:19:50.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title has already been taken."
                                                    }
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The description has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}